fs: Move to integration tests
1.2s for test build -> 0.6s
