+++ crates/fs/tests/integration/fake_git_repo.rs
@@ -1,0 +1,58 @@
+use fs::{FakeFs, Fs};
+use gpui::BackgroundExecutor;
+use serde_json::json;
+use std::path::Path;
+use util::path;
+
+#[gpui::test]
+async fn test_checkpoints(executor: BackgroundExecutor) {
+    let fs = FakeFs::new(executor);
+    fs.insert_tree(
+        path!("/"),
+        json!({
+            "bar": {
+                "baz": "qux"
+            },
+            "foo": {
+                ".git": {},
+                "a": "lorem",
+                "b": "ipsum",
+            },
+        }),
+    )
+    .await;
+    fs.with_git_state(Path::new("/foo/.git"), true, |_git| {})
+        .unwrap();
+    let repository = fs
+        .open_repo(Path::new("/foo/.git"), Some("git".as_ref()))
+        .unwrap();
+
+    let checkpoint_1 = repository.checkpoint().await.unwrap();
+    fs.write(Path::new("/foo/b"), b"IPSUM").await.unwrap();
+    fs.write(Path::new("/foo/c"), b"dolor").await.unwrap();
+    let checkpoint_2 = repository.checkpoint().await.unwrap();
+    let checkpoint_3 = repository.checkpoint().await.unwrap();
+
+    assert!(
+        repository
+            .compare_checkpoints(checkpoint_2.clone(), checkpoint_3.clone())
+            .await
+            .unwrap()
+    );
+    assert!(
+        !repository
+            .compare_checkpoints(checkpoint_1.clone(), checkpoint_2.clone())
+            .await
+            .unwrap()
+    );
+
+    repository.restore_checkpoint(checkpoint_1).await.unwrap();
+    assert_eq!(
+        fs.files_with_contents(Path::new("")),
+        [
+            (Path::new(path!("/bar/baz")).into(), b"qux".into()),
+            (Path::new(path!("/foo/a")).into(), b"lorem".into()),
+            (Path::new(path!("/foo/b")).into(), b"ipsum".into())
+        ]
+    );
+}
