+++ crates/fs/tests/integration/main.rs
@@ -1,0 +1,2 @@
+mod fake_git_repo;
+mod fs;
